IA-3039 Refactor: Split entities/entity_types viewset
